backend/requirements-minimal.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
pytest==7.4.3
python-multipart==0.0.6
pyyaml==6.0.1